Fixed the delete endpoint so that it checks if you are in /orders

+++ public/rest/cus/orders.php
@@ -55,9 +55,8 @@ switch ($method){
             return;
         }
     case "DELETE":
-        if (array_key_exists(4, $pathParts) && is_numeric($pathParts[4])){
+        if (array_key_exists(4, $pathParts) && is_numeric($pathParts[4]) && array_key_exists(3, $pathParts) && $pathParts[3] == 'orders'){
             $orderModel->deleteOrder($pathParts[4]);
-            http_response_code(200);
             return;
         } else {
             http_response_code(400);
